Add files via upload
Importing data from the backup logs and printing it to the log template has successfully been implemented. Excel templates will need to be formatted to properly display the data from the backup logs and the graph. The program takes several minutes to process and display the information onto the new log. An attempt will be made to streamline the code for better performance, but the project will continue in this state should optimization fail.